Extract PoseEstimate into its own file.
This is to remove the dependency of GlobalTrajectoryBuilderInterface
on the TrajectoryBuilder.
