+++ cartographer/mapping_3d/local_trajectory_builder.h
@@ -34,12 +34,10 @@
 namespace cartographer {
 namespace mapping_3d {
 
-// Wires up the local SLAM stack (i.e. UKF, scan matching, etc.) without loop
-// closure.
+// Wires up the local SLAM stack (i.e. pose extrapolator, scan matching, etc.)
+// without loop closure.
 class LocalTrajectoryBuilder {
  public:
-  using PoseEstimate = mapping::GlobalTrajectoryBuilderInterface::PoseEstimate;
-
   struct InsertionResult {
     common::Time time;
     sensor::RangeData range_data_in_tracking;
@@ -60,7 +58,7 @@ class LocalTrajectoryBuilder {
       const sensor::PointCloud& ranges);
   void AddOdometerData(common::Time time,
                        const transform::Rigid3d& odometer_pose);
-  const PoseEstimate& pose_estimate() const;
+  const mapping::PoseEstimate& pose_estimate() const;
 
  private:
   std::unique_ptr<InsertionResult> AddAccumulatedRangeData(
@@ -73,7 +71,7 @@ class LocalTrajectoryBuilder {
   const proto::LocalTrajectoryBuilderOptions options_;
   ActiveSubmaps active_submaps_;
 
-  PoseEstimate last_pose_estimate_;
+  mapping::PoseEstimate last_pose_estimate_;
 
   MotionFilter motion_filter_;
   std::unique_ptr<scan_matching::RealTimeCorrelativeScanMatcher>
